requirements.txt
anaconda==0.0.1.1
arcgis==1.5.2.post1
archook==1.2.0
astroid==2.0.4
autopep8==1.4.4
backcall==0.1.0
bleach==3.1.0
colorama==0.4.0
cycler==0.10.0
decorator==4.3.2
defusedxml==0.5.0
Django==2.1.7
django-cors-headers==3.1.0
django-crispy-forms==1.7.2
django-dms==0.1.1b0
django-filter==2.1.0
django-pyodbc==1.1.3
django-tables2==2.0.4
django-webpack-loader==0.6.0
djangorestframework==3.10.2
entrypoints==0.3
et-xmlfile==1.0.1
GDAL==2.3.3
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.4
jdcal==1.4.1
jedi==0.13.2
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.2.4
jupyter-core==4.4.0
jupyterlab==0.35.4
jupyterlab-server==0.2.0
keyring==17.1.1
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
matplotlib==3.0.2
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.4.0
nbformat==4.4.0
notebook==5.7.4
numpy==1.16.1
openpyxl==2.6.2
pandas==0.24.0
pandocfilters==1.4.2
parso==0.3.2
pbr==5.1.3
pickleshare==0.7.5
Pillow==5.3.0
prometheus-client==0.5.0
prompt-toolkit==2.0.8
psycopg2==2.7.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.3
pycodestyle==2.5.0
Pygments==2.3.1
pyldap==2.4.45
pylint==2.1.1
pyodbc==4.0.24
pyparsing==2.3.1
pyproj==1.9.6
pypyodbc==1.3.4
pyshp==2.0.1
python-dateutil==2.7.5
pytz==2018.5
pywin32-ctypes==0.2.0
pywinpty==0.5.5
pyzmq==17.1.2
Send2Trash==1.5.0
Shapely==1.6.4.post1
six==1.11.0
stevedore==1.30.1
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
virtualenv==16.4.3
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
winkerberos==0.7.0
wrapt==1.10.11
